requirements.txt
flask
flask-cors
pandas
duckdb
python-dotenv
requests
openai
